pyproject.toml
[project]
name = "wowaudit-simple-bot"
version = "0.1.0"
description = "Raid readiness dashboard for WoW guilds, backed by wowaudit team data."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[project.scripts]
wowaudit-check = "wowaudit_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wowaudit_bot"]
